% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment_list.R
\name{attachment.list}
\alias{attachment.list}
\title{List attachments of an experiment}
\usage{
attachment.list(site, usr, pwd, experiment_id)
}
\description{
List attachments of an experiment
}
\details{
\code{List attachment files of an experiment from Cytobank}.
}
\examples{
\donttest{# List attachments of an experiment from community server:
community="https://community.cytobank.org/cytobank"
attachments=attachment.list(community, username, password, 41956)
}
}
