% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_dataframe.R
\name{cyto_dataframe}
\alias{cyto_dataframe}
\title{Convert a parsed list -> Cytobank dataframe}
\usage{
cyto_dataframe(parsed_list)
}
\description{
Convert a parsed list -> Cytobank dataframe
}
\details{
\code{cyto_dataframe}.
}
\examples{
\donttest{# Convert parsed list
df=cyto_dataframe(parsed.list)
}
}
