% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_check.R
\name{dir.check}
\alias{dir.check}
\title{Check if a directory is existing. If not generate it.}
\usage{
dir.check(path)
}
\description{
Check if a directory is existing. If not generate it.
}
\details{
\code{Generates a directory if it is not existing}.
}
\examples{
\donttest{# Generate diroctory, returns path:
path=dir.check("path/to/store/files")
}
}
