% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_file_join.R
\name{directory_file_join}
\alias{directory_file_join}
\title{File path helper function -Ensure path ends in "directory separator"}
\usage{
directory_file_join(directory_path, file_name)
}
\description{
File path helper function -Ensure path ends in "directory separator"
}
\details{
\code{Join a file to a path}.
}
\examples{
\donttest{directory_file_join(directory_path, file_name)
}
}
