% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_download.R
\name{fcs.download}
\alias{fcs.download}
\title{Download several or all fcs-files of an experiment}
\usage{
fcs.download(site, usr, pwd, experiment_id, fcs_file_ids, filenames,
  directory = getwd())
}
\description{
Download several or all fcs-files of an experiment
}
\details{
\code{Download fcs-files of interest}.
}
\examples{
\donttest{# Download fcs-files from community server
community="https://community.cytobank.org/cytobank"
files=fcs.list(community, username, password, 45682)
fcs.download(community, username, password, files$ID[1], 
files$id, files$filename, directory="path/to/store/files")
}
}
