% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_list.R
\name{fcs.list}
\alias{fcs.list}
\title{Get FCS-Files list for an experiment}
\usage{
fcs.list(site, usr, pwd, experiment_id)
}
\description{
Get FCS-Files list for an experiment
}
\details{
\code{fcs.list}.
}
\examples{
\donttest{# Get fcs-files from cytobank community server
community="https://community.cytobank.org/cytobank"
fcs.files=fcs.list(community, username, password, 60068)
}
}
