% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_experiments.R
\name{get.experiments}
\alias{get.experiments}
\title{Get all public experiments from community or premium cytobank, returns dataframe}
\usage{
get.experiments(site, usr, pwd)
}
\description{
Get all public experiments from community or premium cytobank, returns dataframe
}
\details{
\code{get.experiments} .
}
\examples{
\donttest{# Download experiment list
community="https://community.cytobank.org/cytobank"
experiments=get.experiments(community, username, password)
}
}
