% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logIn.R
\name{logIn}
\alias{logIn}
\title{Login to community or premium cytobank server}
\usage{
logIn(site, usr, pwd)
}
\description{
Login to community or premium cytobank server
}
\details{
\code{logIn returns a session object} .
}
\examples{
\donttest{# Login to community server
community="https://community.cytobank.org/cytobank"
session=logIn(community,username,password)
}
}
