% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_download.R
\name{project.download}
\alias{project.download}
\title{Download all or several experiments of a Cytobank project}
\usage{
project.download(site, usr, pwd, df, directory = getwd())
}
\description{
Download all or several experiments of a Cytobank project
}
\details{
\code{Download a whole project from Cytobank}.
}
\examples{
\donttest{# Download project Spitzer:
community="https://community.cytobank.org/cytobank"
experiments=get.experiments(community)
### Filter experiments by keyword
spitzer=search.key(experiments,"Spitzer")
project.download(community, username, password, spitzer,"path/to/download/file")
}
}
