% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_key.R
\name{search.key}
\alias{search.key}
\title{Search keyword in experiment list}
\usage{
search.key(df, keyword = "Spitzer")
}
\description{
Search keyword in experiment list
}
\details{
\code{search a keyword}.
}
\examples{
\donttest{# Search for Spitzer in experiment list
result=search.key(experiments, "Spitzer")
}
}
